stuff stuff stuff

+++ src/Objects/Node.cpp
@@ -7,7 +7,7 @@ void Node::AddChild(Node* c)
 
 void Node::UpdateChild()
 {
-	for (int i = 0; i < m_childNode.size(); i++)
+	for (int i = 0; i < (int)m_childNode.size(); i++)
 	{
 		m_childNode[i]->Transform.Position = Transform.Position;
 		m_childNode[i]->Transform.Angle = Transform.Angle;
